Final epic for The Pledge
