+++ static/src/javascripts/projects/common/modules/experiments/ab-tests.js
@@ -3,8 +3,8 @@ import { commercialPrebidSafeframe } from 'common/modules/experiments/tests/comm
 import { commercialCmpUiIab } from 'common/modules/experiments/tests/commercial-cmp-ui-iab';
 import { askFourEarning } from 'common/modules/experiments/tests/contributions-epic-ask-four-earning';
 import { articlesViewed } from 'common/modules/experiments/tests/contributions-epic-articles-viewed';
-import { articlesViewedMomentVariant } from 'common/modules/experiments/tests/contributions-epic-articles-viewed-moment-position';
-import { articlesViewedMoment60DaysVariant } from 'common/modules/experiments/tests/contributions-epic-articles-viewed-moment-60-days';
+import { articlesViewedMonthMomentFinal } from 'common/modules/experiments/tests/contributions-epic-articles-viewed-month-moment';
+import { articlesViewed60DaysMomentFinal } from 'common/modules/experiments/tests/contributions-epic-articles-viewed-60-days-moment';
 import { countryName } from 'common/modules/experiments/tests/contributions-epic-country-name';
 import { acquisitionsEpicAlwaysAskIfTagged } from 'common/modules/experiments/tests/acquisitions-epic-always-ask-if-tagged';
 import { adblockTest } from 'common/modules/experiments/tests/adblock-ask';
@@ -28,8 +28,8 @@ export const concurrentTests: $ReadOnlyArray<ABTest> = [
 ];
 
 export const epicTests: $ReadOnlyArray<EpicABTest> = [
-    articlesViewedMomentVariant,
-    articlesViewedMoment60DaysVariant,
+    articlesViewedMonthMomentFinal,
+    articlesViewed60DaysMomentFinal,
     articlesViewed,
     learnMore,
     countryName,
